fixed a long-standing API wart regarding the "backing up" behavior.
prior to this commit, the raw incremental interface cannot refer to
the problematic byte not in the current input. this normally affects
the strict incremental usages only, but sometimes (most evidently,
in ISO 2022-JP) it also affects the one-shot usages.

this commit allows `upto` field in the `CodecError` struct to be
negative, and updates the existing definition of `Encoding::{decode,
encode}` to handle this situation correctly.