moved `index` and its submodules into individual crates.
`encoding::index` module was semi-internal but still exposed;
this commit moves it to individual crates on which the main
`encoding` crate depends. they can be used without the main crate.

`encoding-index-*` crates are individually versioned according to
the Encoding Standard; any changes to the identifier would update
the version. for the monotonicity the crate version is based on
the date field, but it doesn't really affect the crate version.