0.2.8: updated indices and changes to euc_kr.
This updates to indices from the Encoding standard 2014-11-06.
Most indices remains unchanged but have been intentionally updated
so that the subpackages for older Rust do not remain.

Importantly, both an index and algorithm for euc_kr has been changed
so that the error handling algorithm is more similar to that of Big5.
The version of the Korean index subpackage has been bumped to reflect
this important change (so be aware of any incompatibility).

I've also found a bug from the table, which have been temporarily
worked around. I'll pull this workaround as soon as the bug is resolved.