0.2.31: partially caught the current WHATWG standard up.
HZ has been removed (only from the WHATWG labels), and GBK is
separated from GB 18030 due to the interoperability problem.
This still lacks the up-to-date ISO-2022-JP encoding which needs
more work, so be careful.

This version also removes long-ignored stability markers that are
to be deprecated on user libraries. This bumps subpackage versions.